refactor(RacingTest):  RacingTest 분리
Racing 관련 테스트가 너무 많아지는 것 같아서 메서드 별로 분리함
