+++ src/test/java/racingcar/RacingInitParticipantsTest.java
@@ -1,15 +1,15 @@
 package racingcar;
 
 import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-public class RacingTest {
+public class RacingInitParticipantsTest {
 
     private static Racing racing;
 
@@ -84,53 +84,4 @@ public class RacingTest {
         }
     }
 
-    
-    
-    @Test
-    @DisplayName("initPlayCount 부족 실패 테스트")
-    void initPlayCount_lack_fail() {
-
-        assertSimpleTest(() ->
-                assertThatThrownBy(() -> racing.initPlayCount(0))
-                        .isInstanceOf(IllegalArgumentException.class)
-        );
-    }
-
-    @Test
-    @DisplayName("initPlayCount 초과 실패 테스트")
-    void initPlayCount_over_fail() {
-
-        assertSimpleTest(() ->
-                assertThatThrownBy(() -> racing.initPlayCount(11))
-                        .isInstanceOf(IllegalArgumentException.class)
-        );
-    }
-
-    @Test
-    @DisplayName("initPlayCount 검증 테스트")
-    void initPlayCount_validate() {
-
-        racing.initPlayCount(5);
-        assertThat(racing.getPlayCount()).isEqualTo(5);
-    }
-
-
-
-    @Test
-    @DisplayName("getPlayResults 검증 테스트")
-    void getPlayResults_validate() {
-
-        String[] carNames = {"디", "레옹", "마틸다", "로즈", "고든"};
-        int playCount = 5;
-
-        racing.initParticipants(carNames);
-        racing.initPlayCount(playCount);
-
-        String result = String.join(
-                "\n",
-                racing.getPlayResults()
-        );
-
-        assertThat(result).contains("디 : ", "레옹 : ", "마틸다 : ", "로즈 : ", "고든 : ", "-");
-    }
 }
